\name{heigh_propotion}
\alias{heigh_propotion}
\title{Cumulative denseness for each line}
\usage{
heigh_propotion(imagematrix)
}
\arguments{
\item{imagematrix}{The matrix to be analysed.}
}
\description{
Proportion of black pixel below each matrix line.
}
\examples{
# First, get a matrix from your image. Here an example of a bush image is used.
bush<-system.file("extdata/bush.JPG",package ="bwimage")
bush_imagematrix<-threshold_color(bush,  "jpeg",  "proportional", compress_rate = 0.1)

# Proportion of black pixel below each matrix line.
heigh_propotion(bush_imagematrix)
}
\references{
Zehm et al 2003 Multiparameter analysis of vertical vegetation structure based on digital image processing. Flora-Morphology, Distribution, Functional Ecology of Plants, 198: 142-160.
}
\seealso{
threshold_color
}
\author{
Carlos Biagolini-Jr.
}
