% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageAnalysis.R
\name{hole_section_data}
\alias{hole_section_data}
\title{Summary of holes information}
\usage{
hole_section_data(section, color = 0)
}
\arguments{
\item{section}{Section to be analysed.}

\item{color}{Color of the hole (0 or 1).}
}
\value{
Number, mean, standard deviation, minimum and maximum size of hole sections.
}
\description{
Summary information of holes of a given color in a given section.
}
\details{
Description of the bush image example provided by bwimage package: The imagem comprises a vegetation section of 30x100cm. To get this photo, in one of the 100cm side of the panel of 100x100cm covered with a white cloth was placed perpendicular to the ground. A plastic canvas of 50x100cm was used to cover the vegetation along a narrow strip in front of a camera positioned on a tripod at a height of 55 cm. A photograph of the portion of standing vegetation against the white cloth was taken.
}
\examples{
# First, get a matrix for your image. Here an example of a bush image is used.
bush<-system.file("extdata/bush.JPG",package ="bwimage")
bush_imagematrix<-threshold_color(bush,  "jpeg",  "proportional", compress_rate = 0.1)
# For further information, see : help(threshold_color)

# Detail information of white (0) holes sections in the column 200 of bush image
hole_section_data(bush_imagematrix[,200], color = 0)

# Detail information of black (1) holes sections in the row 250 of bush image
 hole_section_data(bush_imagematrix[250,], color = 1)

 # See also
 help(hole_section)

}
\references{
Zehm et al 2003 Multiparameter analysis of vertical vegetation structure based on digital image processing. Flora-Morphology, Distribution, Functional Ecology of Plants, 198: 142-160.
}
