% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageAnalysis.R
\name{light_gap}
\alias{light_gap}
\title{Light gap}
\usage{
light_gap(imagematrix, width_size = NA, scale = TRUE)
}
\arguments{
\item{imagematrix}{The matrix to be analysed.}

\item{width_size}{Real size of image width (in mm, cm, m, etc..).}

\item{scale}{If FALSE do not ajust the output for real size.}
}
\value{
Distances without black pixel in each side of the picture

denseness, mean and standard deviations
}
\description{
Left and right distances from first black pixel to image edge.
}
\details{
Description of the bush image example provided by bwimage package: The imagem comprises a vegetation section of 30x100cm. To get this photo, in one of the 100cm side of the panel of 100x100cm covered with a white cloth was placed perpendicular to the ground. A plastic canvas of 50x100cm was used to cover the vegetation along a narrow strip in front of a camera positioned on a tripod at a height of 55 cm. A photograph of the portion of standing vegetation against the white cloth was taken.
}
\examples{
# First, get a matrix for your image. Here an example of a bush image is used.
bush<-system.file("extdata/bush.JPG",package ="bwimage")
bush_imagematrix<-threshold_color(bush,  "jpeg",  "proportional", compress_rate = 0.1)
# For further information, see : help(threshold_color)

# Calculate vegetation Light gap in the bush image matrix
light_gap(bush_imagematrix,width_size=100)
}
\references{
Zehm et al 2003 Multiparameter analysis of vertical vegetation structure based on digital image processing. Flora-Morphology, Distribution, Functional Ecology of Plants, 198: 142-160.
}
