% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvIM.R
\name{stretch}
\alias{stretch}
\title{stretch circle to square}
\usage{
stretch(imagematrix, stretch_method = "radial")
}
\arguments{
\item{imagematrix}{The matrix to be stretched.}

\item{stretch_method}{Stretch algorithm. Four algorithms (radial, shirley, squircle, and elliptical) are available to stretch the image. The algorithms were adapted from Lambers 2016.}
}
\value{
A matrix of 0, 1 and NA representing white, black and transparent pixels, respectively.
}
\description{
Stretch data data from circular image to square in binary matrix
}
\examples{
img_location <- system.file("extdata/chesstable.png",package ="bwimage")
circular_matrix<- threshold_color(img_location,"png")
stretched_matrix<-stretch(circular_matrix,stretch_method="radial")
}
\references{
Lambers 2016 Mappings between Sphere, Disc, and Square. Journal of Computer Graphics Techniques, 5(2): 1-21.
}
\author{
Carlos Biagolini-Jr.
}
