% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImageAnalysis.R
\name{topline}
\alias{topline}
\title{Top line}
\usage{
topline(imagematrix, height_size = NA, width_size = NA)
}
\arguments{
\item{imagematrix}{The matrix to be analysed.}

\item{height_size}{Real size of image height (in mm, cm, m, etc..).}

\item{width_size}{Real size of image width (in mm, cm, m, etc..).}
}
\value{
Line size that cover black pixels
}
\description{
Line running along the crest of highest black pixel.
}
\details{
Description of the bush image example provided by bwimage package: The imagem comprises a vegetation section of 30x100cm. To get this photo, in one of the 100cm side of the panel of 100x100cm covered with a white cloth was placed perpendicular to the ground. A plastic canvas of 50x100cm was used to cover the vegetation along a narrow strip in front of a camera positioned on a tripod at a height of 55 cm. A photograph of the portion of standing vegetation against the white cloth was taken.
}
\examples{
# First, get a matrix from your image. Here an example of a bush image is used.
bush<-system.file("extdata/bush.JPG",package ="bwimage")
bush_imagematrix<-threshold_color(bush,  "jpeg",  "proportional", compress_rate = 0.1)

# See the proportion of black pixels (1) below each bush image matrix row
topline(bush_imagematrix,100,100)
# Conclusion: topline size is 785.6 cm.
}
\references{
Zehm et al 2003 Multiparameter analysis of vertical vegetation structure based on digital image processing. Flora-Morphology, Distribution, Functional Ecology of Plants, 198: 142-160.
}
